Fix bug 697323, Segfault after pattern with transparency.
While the pdfwrite device usage may have contributed to the path being
taken, the segfault was in the GC and was due to a pattern-clist device
still existing in the allocator, but no longer referenced (at least as
seen with a debug build. The fix was to set gx_device_retain false for
the adev (pattern accum device, which may be a pattern-clist device) for
the case where the pattern uses transparency so the freeing the saved
gstate would free the pdf14 device and it would rc_decrement its target,
thus freeing the accumulator device.