Fix 'corner' radial gradient case.
Bug #697144 " Some incorrectly rendered radial shadings"

This is to address 'radialshade4.ps, but fixed (the customer had managed
to create incorrect PostScript for this specific test, see radialshade4b.ps).

In the case where the two circles defining a radial gradient are
tangential, and the larger circle has 'extend' set to true, the extended
area should be limited to the tangent which touches the two circles, and
the smaller circle should not be filled.

Getting this corre4ct with our existing method for drawing gradients
has been difficult. It 'looks like' all gradients are defined in terms
of tensor patches, rather than having a dedicated plotter for each
gradient type.

In order to generate a proper 'truncated' annulus for this case, it is
necessary to construct a number of different 'quadrilaterals' and fit
them carefully around the larger circle. To ensure no 'cracks' appear
between the quads, we must also take care to ensure that they join at
horizontal or vertical boundaries (in user space).

There's some trigonometry involved to get this correct, and to avoid
having to check for infinity returns from tan(alpha), and generally
avoid the complex work required in the general case, we handle the
cases where the tangent point of the circles is on a horizontal or
vertical axis.

This passes cluster tests, the customer test suite, and my own hand
created PostScript files which exercise the code rather more thoroughly

No differences expected.