pdfwrite - fix calculation of a bounding box
Bug #697237 "ps2pdf not producing correct PDF with a rotated character in the PS"

The code for detecting a glyph which is completely clipped out of the
output had a fault; after calculating the Font BoundingBox in current
user space, it only considered three of the points of the resulting
rectangle to calculate the largest rectangle.

This erroneous rectangle was then used to decide if a glyph was entirely
outside the clipping area. Since the top left corner was not considered
a sufficiently large glyph rotated to an acute enough anti-clockwise
angle could appear to be totally clipped when in fact it is only
partially clipped.

Fix this here by considering all four points.

This show two very minor progressions in Quality Logic tests and some
PXL->PDF conversions are now larger. The reason is that we use the
font bounding box and a couple of glyphs are now considered to be
unclipped when in fact they are totally clipped out, but we can't tell.
This results in not only the glyphs but also a font being embedded
in the output and increasing the size slightly.