ps2write - don't try to alter /pagesave when modifying media size
Bug #697245 " Generated postscript by ps2write causes limitcheck for save error"

When altering the media size for a page the code defined the save object
'pagesave' with the new media size. I believe the intention was probably
to have the new media size persist, in order to avoid calling setpagdevice
for every subsequent page with the modified media.

Unfortunately, this does not remove the prior 'save' object attached to
/pagesave. The only way to get rid of that is to actually 'restore' it.

So, fundamentally, we need to remove the definition of pagesave from
the code which sets the media size. We can achieve the same effect in a
rather more roundabout fashion by also moving the definition of pagesave
within the page stream to a point after the definition of the page
dictionary.

Less neat, but it works.

For some obscure reason this results in some pages being rendered
very slightly different sizes on the cluster. I've checked and the
current sizes are correct. I have no idea why this was wrong before
and I'm not interested enough to go looking, since they are now
correct.