PDF interpreter - when detecting transparent pages, include Form 'Groups'
Bug #697354 "PDF Interpreter fails to push pdf14 compositor"

We have an optimisation in the PDF interpreter which ignores page
level Group definitions, as these have no effect on the output.

However, the way this is performed meant that we also did not detect
Group entries in Form XObjects. So if the only transparency on a page
was a Form XObject with a Group entry, then we would incorrectly
decide that the page had no transparency.

We do still want to do the performance check at the page level, but its
important that we include Form XObject Groups in the resources check. So
here we specifically check Form XObjects to see if they have a Group.

This produces differences in a few files, surprisingly miniscule
differences. It appears that we have a number of files which include
Forms with Group entries which, as is so often the case with PDF
files produced by naive graphics libraries, don't actually *do* any
real transparency.

Unfortunately this does mean we now process those files slower than
we did before (and ps2write now produces images for them), but I don't
see any real prospect of detecting pointless transparency at this depth
in a PDF file.

So the only differences are small colour changes depending on the
ProcessColorModel of the destination device. None of these appear to
be significant to me.