PDF interpreter - ensure Rect array entries are not indirect objects
Bug #697357 "Can't handle annots where /Rect entries are indirect"

The example file has an insanely constructed Rect entry, where not only
is the /Rect value itself an indirect object, but each of the entries
in the array is also an indirect object.

Here we ensure that the entries in a /Rect for an annotation are
fully resolved.