Fix segv due to improper depth computation
Commit
47294ff5b168d25bfc7db64f51572d64b8ebde91
had an issue that was found when running
eci_altona-test-suite-v2_technical2_x4.pdf
(the Altona file with ALL the pages on one page)
with tiffsep and psdcmyk.  This fixes the segv.
A rending issue remains on this issue.
Bug http://bugs.ghostscript.com/show_bug.cgi?id=697366
was opened to address.