PDF interpreter - skip 'R' operator in invalid context
Bug #697365 "corrupted content stream causes error"

The file has 2 Form XObject whose content streams have a single garbage
byte at the end of the stream, and which forms part of the 'endstream'.

In one case this simply leads to garbage on the stack, which we deal with
already, but in the other case the garbage happens to be a 'R character
which we attempt to process as a token.

Now attempting to process an R operator fails in this case, because the
stack contents are not a pair of integers. However, in the process, it
removes two stack objects which we require for further processing.

There's no way to restore those to the operand stack, and there is no
way to not process the 'R' as a token. We also cannot temporarily
define 'R' as a no-op for the course of a stream, because if the
stream uses named objects which have not already been dereferenced
we need to execute the R operator to resolve them.

This commit modifies the definition of R so that it checks the type
of the expected operands before consuming them. If they are not both
integers then it flags an error, leaves the stack alone and exits.

THis will, of course, be slightly slower than the current approach
which doesn't check types, but it should only be executed once
for each object in the file. Even for a large file that should only
be a few tens of thousands, and this is likely to be lost among
the processing time for real operations. However, shuold this
causew significant performance problems we may remove this in the
future.

No differences expected.