Revert commit 3cde6d6d, require %d OutputFile spec for multi-page tiffsep
The previous commit allowed multi-page separation files, but this would only
work if all pages had the same spot colors. If a later page added a spot color
it would be appended to the set of spot colors, but then the separation layer
would be emitted with the previous spot color name (according to the order)
and the CMYK equivalent for that previous spot color would be used to make the
CMYK composite.

Now multi-page tiffsep requires the use of %d (or some other format that uses
the page number in the filename). The first page can be output without %d, and
the second page will issue an error message and quit with ioerror.

In theory we could wait to issue the error until we encounter a spot color order
that was not the same as the first page, but this is of little utility, and
this change makes tiffsep operate similarly to the psdcmyk device.