pdfwrite - where possible, preserve annotations
This allows us to preserve many (but not all) annotation from an
input PDF file as annotations in the output PDF file. This only
works with the pdfwrite device, the ps2write device does not
write pdfmarks to the output.

Certain kinds of annotations (eg Link annotations) cannot be
preserved, because they require a page (or other object) destination
which we can't easily know at the time we create the pdfmark.

Widget annotations are not preserved as these only have relvance
in AcroForms, and since we don't preserve AcroForms, its much
better to render the Widgets, otherwise their appearance will
be lost along with the AcroForm data.

It is possible that this behaviour may not be desirable, and
that there could be bugs in this code, so a new switch
'-dPreserveAnnots' is defined for use with the pdfwrite device
only. If set to false the old behaviour returns, and annotations
will be rendered instead of preserved.

There are a few differences with this code; a few progressions,
a few 'differences' and one file where the Stamp annotation no
longer renders. THis is because the Form for the Apperance is
invliad (leaves junk on the stack) and attempts to work around
this ran into al kinds of trouble. In the end I ran out of time
and chose to accept the error, the form is incorrect after all.