PDF Interpreter - have warning messages respect QUIET
Bug #697394 "stderr is included with stdout when result is set to stdout"

We do want the warnings to go to stdout, for debugging purposes,
because it is easier to use -dPDFDEBUG if the warnings are interleaved.

For the benefit of user piping the output to stdout, have the warning
messages respect -dQUIET so they don't end up in the output file.

Error messages (which are directed to stderr anyway) are not suppressed
as these are important.

No differences expected