pdfwrite - adjust Mono Subsample resolution to integer
Bug #697351 "/Subsample filter needs to make non-integer scales integer for monochrome images"

The only downsampling filter available for monochrome images and
imagemasks (1 bpp images and masks) is the /Subsample filter. However
the Subsample filter only works when the downsampling factor is an
exact integer.

Its unreasonable to insist that all the monochrome images in a given
document can be downsampled to a specific resolution and that the downsample
factor for all those images will be an integer.

So here, if the image is monochrome (so we canot switch to the Bicubic
filter) and the factor is not an integer, we force the factor to the
nearest  integer.

Also, update the documentation which incorrectly stated that when
/PDFSETTINGS was set to 'prepress', monochrome images used the Bicubic
filter, it actually uses the Subsample filter.