txtwrite - sort out endian-ness of Unicode based on architecture
Bug #697339 "Device txtwrite seems to be broken"

decode_glyph sppears always to return big-endian shorts for Unicode
code points. On litlee-endian platforms, reverse the byte order so
that we can treat the data as shorts instead of bytes.

As I don't have a big endian device to test this on, further work may
be required.