PDF Interpreter - more indirect object foolishness
Bug #697402 "PDF rendered as blank in PDF with unusual font width array"

Not ony the /Widths array of one of the fonts, but also an entry in a
/W array for a CIDFont specify a value using an indirect object
reference.

While this is silly (it makes the PDF file larger and slower) it is legal
so this commit copes with it by dereferencing any indirect references.

We alos now handle W2 arrays with references, even though this file
doesn't use one.