Bug 697415: 'clean up' after images with SMask entries
Our handling of images with SMasks depends on telling the transparency
compositor to expect an SMask image, draw the SMask as a regular image,
tell the compositor the SMask is done, and then draw the 'main' image (to
which the compositor will apply the SMask).

The problem was that the compositor treats SMasks from images and SMasks from
ExtGState as the same (which they really are), but the image drawing code
took no action to inform the compositor we were done with the SMask, thus
leaving it place.

By adding a call to "unset" the SMask from the current graphics state (in this
case, it's really the current group in the compositor), we ensure the SMask
is only applied to the current image).