Bug 693307 Overprint issues
Remove the portions of code used to simulate
the overprinting of CMYK and spot colorants while
in an RGB device.  This really can't work and
we are doing a real separation compositor device for
this process.  The code parts removed only confuse
understanding of the overprint compositor.  Also fix
problems related to Bug 693307 which were caused by
some confusion in the code with regard to the overprint
mode.  Fix other issues related to the Ghent overprint
tests where we were not handling properly the case
of overprinting with a gray color when the output
device was CMYK based.  Finally there were multiple
issues with the testing and setting of the overprint
settings and the use of the effective overprint mode