Bug 697435 add blending color space
This adds the ability to specify the default (based) blending
color space for a target device when we have transparency blending.
-sBlendColorProfile="my_profile.icc"  is used for the specification.
Only Gray, RGB, or CMYK ICC profiles are allowed.  Also, separation
devices (e.g. tiffsep and psdcmyk) will not support the use of this
at this time.  Note also that if a target device has a put_image
procedure where it did its own blending or tag processing, this command
proc is not called if the blending color space is specified as the final color
conversion is applied through begin_typed_image.