Fix accuracy of bbox device with curves in filled paths
Bug #697446 "bbox device is inaccurate with certain curves, when filled"

The problem here is that, when a path is not clipped, is composed of
certain types of flattish curves at the boundary, we used an inaccurate
method to determine the bounds of the path.

The old code used gx_path_bbox, which only considers the points in a path
which is a problem for curves, because it considers the control points
to be part of the curve, which they are not, and with some curves
they can lie a long way from the actual path, as in this case.

This commit simply forces the code through the scan-converter in order
to get an accurate result, the same as the stroked code.

No differences expected.