pdfwrite - don't emit degenerate Matrix for type 2 Patterns
Bug 697451 "shfill with degenerate matrix leads to PDF unreadable with Acrobat"

The original file deliberately makes the CTM degenerate before drawing
a shfill. Although all PostScript consumers handle this without complaint.
When converted to PDF most PDF consumers also are happy with the
situation, however Adobe Acrobat recoils in horror and aborts the
processing of the page stream with an error.

Adobe Acrobat Distiller simply refuses (silently!) to embed the shfill
in the PDF file.

We don't have a way to drop the fill, so instead, if the Matrix is
degenerate, replace it with the smallest scale matrix we can. Acrobat
is happy with this.

No differences expected.