New scan converter: Fix problems seen at local minima/maxima.
Sometimes the fill isn't quite right at the extremes of shapes.

This is due to a problem when we exactly align with scanlines.

See the output of tests_private/comparefiles/Bug696174.ps at 300dpi
for some examples. Look at the tie lines for the first notes in the
last line of page 2 - a scanlines worth of pixels are missing from
the top.