Bug 697456.  Dont create new ctx when pdf14 device reenabled
This bug had yet another weird case where the user created a
file that pushed the pdf14 device twice.  We were in that case,
creating a new ctx and blowing away the original one with out
proper clean up.  To avoid, only create a new one when we need it.