New scan converter: Fix problems in trap fills.
I was calculating trapezium fills slightly wrong.

Also, I was filling the 'centre of a pixel' cases edgebuffers
with slightly incorrect values. Now the debugging shows that
we are getting exactly what we want in the example files I
have tried.

This results in much smoother edges on some shapes.