Bug 697462: pdfwrite: avoid cached glyph use
When pdfwrite is accumulating a Type 3 font, and we want to try to capture the
CharProc (rather than falling back to a bitmap), we need to stop the core
font code from using a previously cached glyph bitmap (for example, from an
earlier stringwidth operation), so we're sure the CharProc gets executed.

If we don't ensure the cache doesn't get used, we end up in an infinite loop,
where pdfwrite repeatedly returns to the core to run the CharProc.