Improve CIDFont rearranged font handling slightly
Noticed while investigating a customer problem. We do not handle
rearranged fonts (because they are a hideous hack) but we do have code
to at least process them. Unfortunately there's a typo in the code which
causes it to throw an error if its ever executed.

Fix the typo here.

The fact that this has never come up indicates how often this mess is
used; ie never.