Fix SEGV bug 697473 bitrgbtags device with transparency
The SEGV was caused by setting tos->n_planes to the wrong value if we
converted colors to a different number of color components. Also fix
clearing of the tags if the backdrop was NULL (leave it set as it was
by pdf14_buf_new to GS_UNKNOWN_TAG).

Lastly, the pdf14_put_image wasn't setting up the buffer pointers for
all of the planes (alpha and tag), so bit_put_image choked. This
probably would have caused a problem with pngalpha.