Bug 697435 bittags device blending color space
In the clist case, there was a mixup with the ICC profile
during playback.  This fixes that issue.  While the small
file in Bug 697433 works with this fix, the large file
(regression3.pdf) still shows 2 issues.  When the blending
color space is -sBlendColorProfile=default_cmyk.icc at 300dpi
we see a dropped band (likely some confusion about the
pdf14/clist optimization).   When we use no blending color
space we end up with a segv in clist_playback_band