Bug 697482: handle cff with broken /Private data
The problem in this case is that the offset into the cff data for the /Private
dictionary definition is well beyond the end of the cff stream.

Bounds check the offset, and if it's nonsense, treat it as a zero length
/Private dictionary (in the hope that the font doesn't actually need it.

NOTE: /Private *is* a required entry, so *using* the font may throw an error.