Add option to hide jpeg entrypoints.
If we define GS_HIDE_INTERNAL_JPEG on build, then a series of macros
is used to rename internal JPEG entrypoints. This can prevent
symbol clashes when linking with other libraries.

There may well be more symbols to add here in future, but this
seems to be enough to solve it for me now.