Force some typedef enums to be 32 bits.
Modern C's are allowed to shrink typedef enums so that they are
only as large as required to cover all the values in the
enumeration. This can mean that typedef enums can now be byte
sized.

This works very badly when we have an array of them, take the
address and cast it to an int *. That int * is no longer
appropriately aligned for some compilers. This happens in the
get_param code.

We therefore add a trailing 'large' enum value to force the
enums to be large. This will break down if we ever have a platform
where ints are larger than 32bits.