Bug 697503: Check resource path, and GenericResourceDir...
...paths when searching for resources.

This relates back to bug 694509.

If GenericResourceDir is set as a relative path, relative to a non-standard
GS_LIB search path, we have to prefix the resource path (Category/Key) with
the GenericResourceDir for the search path machinery to find the resource
file.

*However*, if the GenericResourceDir is set in the normal way (automagically
by the initialization), then just using the plan resource path is required.

So, try both the plain 'Category/Key' path, and the
'GenericResourceDir/Category/Key' path.