Bug 626295 Text Knockout Transparency
When we have transparency, text knockout set,
with either a non-normal blend mode or an
opacity less than 1.0, we need to push a
non-isolated knockout group.  Graphic state
changes can occur between the BT and ET commands.
As such, we will need to handling the push
operation in pdf14_text_begin (and the clist version)
when the conditions are right.  This is a special
group that indicates its uniqueness with a text_group
flag.  This flag is included through the clist, the
group as well as the device.  Once set, subsequent
calls to pdf14_text_begin will not push another group.
The group is popped with ET is encountered through the
zendtransparencytextgroup command from the interpreter.
If the pdf14 device is not currently in a text group,
this group pop is ignored (and not even placed in
the clist).  The pdfwrite compositor logic, ignores
and group pushes that occur with the text_group set
as well as all PDF14_END_TRANS_TEXT_GROUP types.
Due to confusion of with the annotation text writing
methods, I had to add a PDF14_BEGIN_TRANS_TEXT_GROUP
to denote when we encounter a BT in the source file
to ensure that we are only going to push groups if
our conditions are right and we are in a BT/ET pair and
not drawing a text annotation.

One optimization we may want to look at doing later, is
determining the bbox for the area between BT and ET.
Currently the group push is the same size as the parent
group which is likely much to large.