Fix Bug 697502, problems with NumRenderingThreads exceeding available RAM
The logic that determined the amount of printer buffer space, and allocated it
was including the ESTIMATED pdf14 transparency buffer space, but we only need
to make sure that amount of space is available, not actually allocate it (since
then it wouldn't be availble when needed).

if there wasn't enough memory for the band buffer when a thread's clist reader
was opened, gdev_prn_setup_as_command_list would try reducing the BufferSpace
(by successive dividing by 2) until it could fit. This would result in a
reader band size that did not match what was used to write the clist. This
so we force the BandBufferSpace to be exactly the same as was used in the writer
using the cdev page_info.band_params. The tile_cache_size was added to the
space_params used to set up the thread clist device to make it match that used
by the writer.

In order to make sure we don't "over commit" in creating threads we reserve
an extra amount per thread, currently 2 Mb plus the ht_cache size times the
number of components, plus when the page uses transparency, we increase it by
the ESTIMATED_PDF14_ROW_SPACE times the band_height. We further increase it by
the size of the link profiles and the size of the icclink profiles (CURRENTLY
ESTIMATED). We free the reserve area as we start threads.

Also when an allocation for a thread failed, the 'band' used for next_band was
being decremented but it did not need to be since it was already the correct value.
This had caused a problem with threads getting started for the wrong band.

Fix gx_ht_read_and_install to prevent double free if gx_gstate_dev_ht_install
gets an error (VMerror).

Fix gdevbit so that we don't keep going after get_bits returns an error.

TODO: collect the size of the icclinks rather than a rather worst case value
currently estimated.