Memento: Improve memory squeezing.
It seems that Memory squeezing can sometimes run into
problems with the child hanging after a fork. This seems
to be related to the child process SEGVing and the
signal handler not picking it up.

I've introduced a workaround so that the parent now only
waits a maximum of 30 seconds before killing the child and
continuing.