Memento memory squeezing tweaks.
Add Memento_bt() - a function to output the backtrace
to stderr.

Call this when we fork for a memory squeeze.

Forking only clones the active thread, so we can only use
the forking memory squeezer on single threaded tasks. Even
with single threaded tasks, you can't unlock mutexes in
the child that were taken in the parent. This means that
we need to disable mutexes in the main application if we
want to memory squeeze safely.

Also tweak the time we spend sleeping while waiting for a child
to die. This makes a huge difference to the runtime.

If we die during squeezing due to a SEGV, then don't bother
listing the blocks that we leaked.