Tweak gs to enable fast Memento memory squeezing.
Memento can memory squeeze in 2 modes; firstly it can repeatedly
run the same job again and again failing at subsequent points.

This has the advantage of being simple, but it has the problem
of being slow, in that processing is repeated again and again.

A faster mode is available where we run to the point of failure,
then fork. The child proceeds to fail, and once failed, the
parent then continues for 1 more allocation then repeats the
process. Thus the only repeated work is in the failure cases.

The downside to this is that fork copes very badly with
multi-threaded programs. Firstly, it only duplicates the current
thread - this means that we can only use it with single
threaded programs. Secondly, even with single threaded programs
the forked child goes wrong if you try to use mutexes etc.

This means that in order to use GS with the fast memento memory
squeezer, we need to ensure we don't create mutexes etc. We
do this by adding a MEMENTO_SQUEEZE_BUILD define to gs.