XPS interpreter - add a reverse GID lookup for the benefit of pdfwrite
Bug #697304 "Encoding gets lost when generating PDF"

Prior to this commit the XPS interpreter, like the PCL interpreter, did
not really return a character code when asked for an equivalent for a
GID. Instead it returned the 'last' character code.

For strings with multiple characters processed at once, this meant that
all the characters in the string were assigned the same 'code' as the
last character in the string. Also there was no reason to assume that
code was even a Unicode value.

This commit adds code to reverse the TrueType CMAP subtable lookup;
given a GID it will return a character code. We then use the resulting
character code in pdfwrite (which is the only device to use this method)
to build a ToUnicode CMap in the output PDF file.

There are a number of caveats with this commit;

I've only been able to create XPS files with fonts containing format 4
CMAP subtables. While I've coded the others (but not 2 or 8 because the
normal character code -> GID mapping doesn't support them either) I
have no way to actually test whether they work

I'm not certain whether its possible to get an XPS file containing a
font which does not have a 3,1 CMAP. Again I've not been able to create
such a file, if we ever encounter one this code will likely not work
properly.


I have tested Latin and Far Eastern scripts, and files which use more
than 256 glyphs from a single font. All of these appear to work as
expected. When more than 256 glyphs are used from the same font, we
create a PDF file with multiple simple subsets of the font, each with
less than 256 glyphs. Each subset has its own ToUnicode CMap and so
cut&paste works as expected.

No differences expected, we cannot cluster test this code.