Increase PDF 1.4 transparency ESTIMATED_ROW_SPACE
Testing showed that the ESTIMATED_PDF14_ROW_SPACE was too optimistic, so
increase NUM_PDF14_BUFFERS to 4 and use the target num_components (primarily
for spot color devices such as tiffsep) to try and make this more robust.
Still a guess, but works better as tested with multi-threaded rendering.

Note, the band height chosen based on the BufferSpace will be different
when the page has transparency, so there are differences. I have reviewed
them and they are all 1 pixel differences or invisible color differences.