ramfs: fix modes other than (r) and (w)
The (a), (r+) and (w+) modes were not working correctly. The "core" code of the
ramfs was correct, but the "glue" code between the Ghostscript stream API
and the ramfs core was not setting flags to support those.

Conversely, the ramfs "glue" code was not setting the permissions flags
correctly in the stream objects it created, when using the "+" modes.

Also, (w) and (w+) were not truncating the file on opening.

Finally, tweak the definition of the various read/write/etc flags for the
ramfs code so the values match those used in the stream API (this makes fixing
the flags in the stream objects much simpler).