Split out $(MAKE) from recursive make calls
Commit 8acec58309 consolidate the common parts of the various recursive make
calls. Turns out, this breaks parallel make, as it defeats the make
algorithm which spots a make command line. As a result, none of those recursive
make calls would do parallel builds.

So, this time, we keep the options consolidated into a make variable, but put
back the $(MAKE) call explicitly for each recursive call