Fix SAFER issue with filenameforall
With the change to ensure we only apply SAFER permissions to real file system
operations (so only on '%os%' and not on other PS '%<device>%' devices), the
way filenameforall called the permission checking function meant it did not
work correctly.

Basically, check_file_permissions() and co must have a valid iodev parameter
passed to them to function correctly.

So, fix that call.