pdfwrite - don't seg fault for Line Printer font and UFST
Using the PCL Line Printer Font, when built to use UFST, can't be
handled with pdfwrite currently. In fact, using UFST with pdfwrite
results in garbage output anyway.

However, we should not seg fault. So here we check to see if we have a
PDF Font resource associated with a native font, and if we don't have
one when we need it, we give up and signal an invalidfont error.