Fix bug #687561, bad parsing of intellifont data.
The previous code read the metric offset field at the wrong positon (8
vs. 6) resulting in a value that caused a range check error.  Further,
the range checking which checked the offsets were ordered is not
useful or correct, and has been removed.  There is nothing in the
documentation to indicate the offsets must be ordered and showing
order does not help in preventing out of bounds access.