Update bitrgbtags device to use better header.
The bit devices output raw data.

The bittagsrgb device outputs rgb + a tag plane, with a PPM header.
The PPM header tells code to expect 3 bytes of image data, not 4,
so is wrong. Here we change it over to use a PAM header (of type
RGB_TAG).

This allows bmpcmp to be amended to cope better.