Add support for RGB_TAG PAM files to lib/viewpbm.ps (type from bitrgbtags)
The bitrgbtags device was changed recently to create a PAM (P7) file rather
than a bogus P6 file with 4 components. This means that the pcl/tools/GOT
detag.c and tagimage.c no longer worked. Instead of changing these the
lib/viewpbm.ps was enhanced to process the RGB_TAG P7 files.

By default, it shows the image, and with -dTAG it shows pseudo-color for
the tag values, UNKNOWN/UNMARKED is white, TEXT is black, PATH is yellow
and IMAGE is red.

example usage:
   gs -dSCALE=1 -- lib/viewpbm.ps bitrgbags.pam
or
   gs -dSCALE=1 -dTAG -- lib/viewpbm.ps bitrgbags.pam

This can also be used to convert the P7 image to another format using
-sDEVICE=___ -o ___ arguments prior to the -- option.

Also, fix numerous problems with FITPAGE and SCALE options. They were
just WRONG!