Fix SEGV is display device when PageSize needs bitmap > 2Gb
The display_put_params is supposed to revert to the original settings
if an error is encountered, but since it frees the display bitmap before
attempting to allocate the new one, if that allocation fails, the device
is left with a memory device that has an invald bitmap pointer.

Instead revert the settings and attempt to allocate a bitmap with those
settings. If that fails (unlikely) then gs_abort with an error message.

NB: It is preferable to free one bitmap before allocating the new one
in case we are near a system memory constraint.