Fix display device for non-PostScript interpreters
commit cf279b8c4773e14045dd6b411c3a524c6ac39124 fixed a potential
crash when memory was exhausted, but used the gs_abort() routine to
tell the interpreter to abort immediately and not carry on.

Unfortunately, and despite the comment in gsexit.h, gs_abort() is only
available in the PostScript interpreter.

So here we opt to return a fatal error, which *should* cause the
interpreter to give up (it does on PCL and PostScript) and will allow
us to build GhostPCL on Windows.