Return Fatal errors from PS putdeviceparams to avoid potential SEGV
The change to return Fatal from the display device (commit 269354e)
would segfault with the PS interpreter because zputdeviceparams never
returned an error code, assuming that the interpreter could handle it,
but this resulted in a segfault when setpagedevice tried to erase the
page and the display device did not have a valid bitmap. Return the
Fatal error so the interpreter can exit cleanly.without performing
the usual setpagedevice "configurationerror" recovery.